fix(input-parser): 피연산자를 int가 아닌 double로 파싱
이유: 소수가 입력될 수 있기 때문에 피연산자를 double로 파싱

+++ src/main/java/calculator/InputParser.java
@@ -51,7 +51,7 @@ public class InputParser {
 
         for (String str : split) {
             try {
-                Integer.parseInt(str);
+                Double.parseDouble(str);
             } catch(NumberFormatException e) {
                 throw new IllegalArgumentException();
             }
